requirements.txt
# Core framework
streamlit==1.28.0

# Data validation
pydantic==2.5.0

# Environment management
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Type checking
typing-extensions==4.8.0

# Utilities
python-dateutil==2.8.2

